+++ visualization/app/codeCharta/ui/ribbonBar/showScenariosButton/addCustomScenario/addCustomScenario.module.ts
@@ -1,7 +1,7 @@
 import { CommonModule } from "@angular/common"
 import { NgModule } from "@angular/core"
 import { FormsModule, ReactiveFormsModule } from "@angular/forms"
-import { MaterialModule } from "../../../../material/material.module"
+import { MaterialModule } from "../../../../../material/material.module"
 import { AddCustomScenarioComponent } from "./addCustomScenario.component"
 
 @NgModule({